Improve settings design (change background color of box and add blur) + resize box
